refactor: move alerts and derived_streams into a new component called scheduled_ops
