+++ src/common/meta/scheduled_ops/alerts.rs
@@ -1,0 +1,80 @@
+// Copyright 2024 Zinc Labs Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+use config::meta::stream::StreamType;
+use hashbrown::HashMap;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+use super::{QueryCondition, TriggerCondition};
+
+#[derive(Clone, Debug, Serialize, Deserialize, ToSchema)]
+pub struct Alert {
+    #[serde(default)]
+    pub name: String,
+    #[serde(default)]
+    pub org_id: String,
+    #[serde(default)]
+    pub stream_type: StreamType,
+    #[serde(default)]
+    pub stream_name: String,
+    #[serde(default)]
+    pub is_real_time: bool,
+    #[serde(default)]
+    pub query_condition: QueryCondition,
+    #[serde(default)]
+    pub trigger_condition: TriggerCondition,
+    pub destinations: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub context_attributes: Option<HashMap<String, String>>,
+    #[serde(default)]
+    pub row_template: String,
+    #[serde(default)]
+    pub description: String,
+    #[serde(default)]
+    pub enabled: bool,
+    #[serde(default)]
+    /// Timezone offset in minutes.
+    /// The negative secs means the Western Hemisphere
+    pub tz_offset: i32,
+}
+
+impl PartialEq for Alert {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+            && self.stream_type == other.stream_type
+            && self.stream_name == other.stream_name
+    }
+}
+
+impl Default for Alert {
+    fn default() -> Self {
+        Self {
+            name: "".to_string(),
+            org_id: "".to_string(),
+            stream_type: StreamType::default(),
+            stream_name: "".to_string(),
+            is_real_time: false,
+            query_condition: QueryCondition::default(),
+            trigger_condition: TriggerCondition::default(),
+            destinations: vec![],
+            context_attributes: None,
+            row_template: "".to_string(),
+            description: "".to_string(),
+            enabled: false,
+            tz_offset: 0, // UTC
+        }
+    }
+}
