+++ src/handler/http/request/mod.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod alerts;
+pub mod scheduled_ops;
 pub mod authz;
 pub mod clusters;
 pub mod dashboards;
