+++ src/handler/http/request/scheduled_ops/mod.rs
@@ -13,23 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod scheduled_ops;
-pub mod authz;
-pub mod dashboards;
-pub mod functions;
-pub mod http;
-pub mod ingestion;
-pub mod maxmind;
-pub mod middleware_data;
-pub mod organization;
-pub mod pipelines;
-pub mod prom;
-pub mod proxy;
-pub mod saved_view;
-pub mod search;
-pub mod service;
-pub mod stream;
-pub mod syslog;
-pub mod telemetry;
-pub mod traces;
-pub mod user;
+pub mod alerts;
+pub mod derived_streams;
+pub mod destinations;
+pub mod templates;
