+++ src/handler/http/router/mod.rs
@@ -407,24 +407,24 @@ pub fn get_service_routes(cfg: &mut web::ServiceConfig) {
             .service(dashboards::reports::delete_report)
             .service(dashboards::reports::enable_report)
             .service(dashboards::reports::trigger_report)
-            .service(alerts::save_alert)
-            .service(alerts::update_alert)
-            .service(alerts::get_alert)
-            .service(alerts::list_alerts)
-            .service(alerts::list_stream_alerts)
-            .service(alerts::delete_alert)
-            .service(alerts::enable_alert)
-            .service(alerts::trigger_alert)
-            .service(alerts::templates::save_template)
-            .service(alerts::templates::update_template)
-            .service(alerts::templates::get_template)
-            .service(alerts::templates::delete_template)
-            .service(alerts::templates::list_templates)
-            .service(alerts::destinations::save_destination)
-            .service(alerts::destinations::update_destination)
-            .service(alerts::destinations::get_destination)
-            .service(alerts::destinations::list_destinations)
-            .service(alerts::destinations::delete_destination)
+            .service(scheduled_ops::alerts::save_alert)
+            .service(scheduled_ops::alerts::update_alert)
+            .service(scheduled_ops::alerts::get_alert)
+            .service(scheduled_ops::alerts::list_alerts)
+            .service(scheduled_ops::alerts::list_stream_alerts)
+            .service(scheduled_ops::alerts::delete_alert)
+            .service(scheduled_ops::alerts::enable_alert)
+            .service(scheduled_ops::alerts::trigger_alert)
+            .service(scheduled_ops::templates::save_template)
+            .service(scheduled_ops::templates::update_template)
+            .service(scheduled_ops::templates::get_template)
+            .service(scheduled_ops::templates::delete_template)
+            .service(scheduled_ops::templates::list_templates)
+            .service(scheduled_ops::destinations::save_destination)
+            .service(scheduled_ops::destinations::update_destination)
+            .service(scheduled_ops::destinations::get_destination)
+            .service(scheduled_ops::destinations::list_destinations)
+            .service(scheduled_ops::destinations::delete_destination)
             .service(kv::get)
             .service(kv::set)
             .service(kv::delete)
