+++ src/handler/http/router/openapi.rs
@@ -86,24 +86,24 @@ use crate::{common::meta, handler::http::request};
         request::dashboards::folders::get_folder,
         request::dashboards::folders::update_folder,
         request::dashboards::move_dashboard,
-        request::alerts::save_alert,
-        request::alerts::update_alert,
-        request::alerts::list_stream_alerts,
-        request::alerts::list_alerts,
-        request::alerts::get_alert,
-        request::alerts::delete_alert,
-        request::alerts::enable_alert,
-        request::alerts::trigger_alert,
-        request::alerts::templates::list_templates,
-        request::alerts::templates::get_template,
-        request::alerts::templates::save_template,
-        request::alerts::templates::update_template,
-        request::alerts::templates::delete_template,
-        request::alerts::destinations::list_destinations,
-        request::alerts::destinations::get_destination,
-        request::alerts::destinations::save_destination,
-        request::alerts::destinations::update_destination,
-        request::alerts::destinations::delete_destination,
+        request::scheduled_ops::alerts::save_alert,
+        request::scheduled_ops::alerts::update_alert,
+        request::scheduled_ops::alerts::list_stream_alerts,
+        request::scheduled_ops::alerts::list_alerts,
+        request::scheduled_ops::alerts::get_alert,
+        request::scheduled_ops::alerts::delete_alert,
+        request::scheduled_ops::alerts::enable_alert,
+        request::scheduled_ops::alerts::trigger_alert,
+        request::scheduled_ops::templates::list_templates,
+        request::scheduled_ops::templates::get_template,
+        request::scheduled_ops::templates::save_template,
+        request::scheduled_ops::templates::update_template,
+        request::scheduled_ops::templates::delete_template,
+        request::scheduled_ops::destinations::list_destinations,
+        request::scheduled_ops::destinations::get_destination,
+        request::scheduled_ops::destinations::save_destination,
+        request::scheduled_ops::destinations::update_destination,
+        request::scheduled_ops::destinations::delete_destination,
         request::kv::get,
         request::kv::set,
         request::kv::delete,
@@ -168,20 +168,20 @@ use crate::{common::meta, handler::http::request};
             meta::saved_view::DeleteViewResponse,
             meta::saved_view::CreateViewResponse,
             meta::saved_view::UpdateViewRequest,
-            meta::alerts::Alert,
-            meta::alerts::Condition,
-            meta::alerts::Operator,
-            meta::alerts::Aggregation,
-            meta::alerts::AggFunction,
-            meta::alerts::QueryType,
-            meta::alerts::TriggerCondition,
-            meta::alerts::AlertFrequencyType,
-            meta::alerts::QueryCondition,
-            meta::alerts::destinations::Destination,
-            meta::alerts::destinations::DestinationWithTemplate,
-            meta::alerts::destinations::HTTPType,
-            meta::alerts::destinations::DestinationType,
-            meta::alerts::templates::Template,
+            meta::scheduled_ops::alerts::Alert,
+            meta::scheduled_ops::Condition,
+            meta::scheduled_ops::Operator,
+            meta::scheduled_ops::Aggregation,
+            meta::scheduled_ops::AggFunction,
+            meta::scheduled_ops::QueryType,
+            meta::scheduled_ops::TriggerCondition,
+            meta::scheduled_ops::AlertFrequencyType,
+            meta::scheduled_ops::QueryCondition,
+            meta::scheduled_ops::destinations::Destination,
+            meta::scheduled_ops::destinations::DestinationWithTemplate,
+            meta::scheduled_ops::destinations::HTTPType,
+            meta::scheduled_ops::destinations::DestinationType,
+            meta::scheduled_ops::templates::Template,
             meta::functions::Transform,
             meta::functions::FunctionList,
             meta::functions::StreamFunctionsList,
